优化IMU配置和参数类型调整
- 启用EKF算法作为默认IMU解算方式（imu_algorithm_select = 1）
- 修改machine_angle类型：uint32 -> float，支持小数精度调整
- 调整角度保护阈值：1500.0 -> 15.0（更合理的度数范围）
- 优化菜单配置：machine_angle步进改为0.1/1/10/100，显示2位小数
- 代码格式优化：统一注释对齐，提升可读性
- Debug界面优化：pitch角显示从5位整数改为3位+2位小数

+++ code/imu.h
@@ -97,10 +97,10 @@ extern imu_data_t imu_data;
 extern uint8 imu_algorithm_select;
 
 // ---------- IMU校准参数（可在菜单中调整） ----------
-extern uint8 gyro_ration;    // 陀螺仪权重系数（互补滤波用）
-extern uint8 acc_ration;     // 加速度计权重系数（互补滤波用）
-extern float call_cycle;     // 解算周期，单位：秒（互补滤波用）
-extern uint32 machine_angle; // 机械中值角度偏移（度）
+extern uint8 gyro_ration;   // 陀螺仪权重系数（互补滤波用）
+extern uint8 acc_ration;    // 加速度计权重系数（互补滤波用）
+extern float call_cycle;    // 解算周期，单位：秒（互补滤波用）
+extern float machine_angle; // 机械中值角度偏移（度）
 
 // ---------- 陀螺仪零偏校准数据（原始数据） ----------
 extern int16 gyro_x_offset; // X轴零偏
